pi_yo_6/coeiroink/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.7.0
    # via -r requirements.in
anyio==3.3.4
    # via starlette
appdirs==1.4.4
    # via -r requirements.in
asgiref==3.4.1
    # via uvicorn
cffi==1.15.0
    # via soundfile
click==8.0.3
    # via uvicorn
colorama==0.4.4
    # via
    #   click
    #   tqdm
cython==0.29.24
    # via
    #   pyopenjtalk
    #   pyworld
fastapi==0.70.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==3.3
    # via anyio
numpy==1.20.0
    # via
    #   -r requirements.in
    #   pyopenjtalk
    #   pyworld
    #   scipy
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via fastapi
pyopenjtalk @ git+https://github.com/VOICEVOX/pyopenjtalk@a85521a0a0f298f08d9e9b24987b3c77eb4aaff5
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pyworld==0.3.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
scipy==1.7.1
    # via -r requirements.in
six==1.16.0
    # via
    #   pyopenjtalk
    #   python-multipart
sniffio==1.2.0
    # via anyio
soundfile==0.10.3.post1
    # via -r requirements.in
starlette==0.16.0
    # via fastapi
tqdm==4.62.3
    # via pyopenjtalk
typing-extensions==3.10.0.2
    # via pydantic
uvicorn==0.15.0
    # via -r requirements.in
espnet